requirements.txt
numpy
torch
torchvision
Pillow
matplotlib
pandas
scikit-learn
tensorflow
keras
opencv-python
scipy
tqdm 
clip
pytorch-lightning
mlfow
